requirements.txt
aiogram==3.4.1
python-dotenv==1.0.0
pydantic==2.5.3
aiohttp==3.9.1
